controlnet-main/requirements0.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
antlr4-python3-runtime==4.9.3
async-timeout==5.0.1
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
einops==0.8.0
filelock==3.16.1
fonttools==4.55.4
frozenlist==1.5.0
fsspec==2024.12.0
grpcio==1.69.0
hydra-core==1.3.2
idna==3.10
Jinja2==3.1.5
kiwisolver==1.4.8
lightning==2.2.5
lightning-utilities==0.11.2
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
monai==1.3.1
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
nibabel==5.2.1
numpy==1.24.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
opencv-contrib-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.1.0
propcache==0.2.1
protobuf==5.29.3
pydicom==2.4.4
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytorch-lightning==2.3.0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rich==13.7.1
scipy==1.15.1
SimpleITK==2.3.1
six==1.17.0
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
torch==2.5.1
torch-fidelity==0.3.0
torch-tb-profiler==0.4.3
torchmetrics==1.4.0.post0
torchvision==0.20.1
tqdm==4.67.1
triton==3.1.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
Werkzeug==3.1.3
yarl==1.18.3
